% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_fun.R
\name{boot.lucid}
\alias{boot.lucid}
\title{Deprecated function boot.lucid}
\usage{
boot.lucid(G, Z, Y, CoG = NULL, CoY = NULL, model, conf = 0.95, R = 100)
}
\arguments{
\item{G}{Exposures, a numeric vector, matrix, or data frame. Categorical variable 
should be transformed into dummy variables. If a matrix or data frame, rows 
represent observations and columns correspond to variables.}

\item{Z}{Omics data, a numeric matrix or data frame. Rows correspond to observations
and columns correspond to variables.}

\item{Y}{Outcome, a numeric vector. Categorical variable is not allowed. Binary 
outcome should be coded as 0 and 1.}

\item{CoG}{Optional, covariates to be adjusted for estimating the latent cluster.
A numeric vector, matrix or data frame. Categorical variable should be transformed 
into dummy variables.}

\item{CoY}{Optional, covariates to be adjusted for estimating the association 
between latent cluster and the outcome. A numeric vector, matrix or data frame. 
Categorical variable should be transformed into dummy variables.}

\item{model}{A LUCID model fitted by \code{est.lucid}.}

\item{conf}{A numeric scalar between 0 and 1 to specify confidence level(s) 
of the required interval(s).}

\item{R}{An integer to specify number of bootstrap replicates for LUCID model.
If feasible, it is recommended to set R >= 1000.}
}
\description{
This function deprecates. Please use boot_lucid instead.
}
