% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_utils.R
\name{check_imputation_quality}
\alias{check_imputation_quality}
\title{Check quality of imputed data}
\usage{
check_imputation_quality(original, imputed)
}
\arguments{
\item{original}{Original data matrix with missing values}

\item{imputed}{Imputed data matrix}
}
\value{
List containing imputation quality metrics and overall assessment
}
\description{
Check quality of imputed data
}
