% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_K}
\alias{check_K}
\title{Check Validity of 'K'}
\usage{
check_K(K)
}
\arguments{
\item{K}{A vector of integers.}
}
\description{
Checks if the elements in 'K' are integers and greater than or equal to 2.
}
