% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{check_na}
\alias{check_na}
\title{Check missing patterns in omics data}
\usage{
check_na(Z, lucid_model = c("early", "parallel"))
}
\arguments{
\item{Z}{A data matrix or list of matrices for parallel model}

\item{lucid_model}{Model type ("early" or "parallel")}
}
\value{
List containing missing data analysis
}
\description{
Check missing patterns in omics data
}
