% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CovY}
\alias{CovY}
\title{Covariate Set in the X->Y path}
\format{
A set with 3000 rows and 5 variables:
\describe{
  \item{CovY1, CovY2}{Two continuous covariates}
}
}
\usage{
CovY
}
\description{
A simulated dataset containing one of the optional components to run est_lucid,
plot_lucid, and tune_lucid. The variables are as follows:
}
\keyword{datasets}
