% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_lucid.R
\name{def.control}
\alias{def.control}
\title{Control parameters for EM algorithm}
\usage{
def.control(tol = 0.001, max_itr = 1000, max_tot.itr = 10000)
}
\arguments{
\item{tol}{Convergence criteria for the EM algorithm. Default is 0.001.}

\item{max_itr}{Maximum number of iterations in each try of fitting process, integer, default is 1000.}

\item{max_tot.itr}{Maximum number of total iterations, integer, default is 10000.}
}
\value{
A list of tolerance settings for LUCID.
}
\description{
Control parameters for EM algorithm
}
\author{
Yinqi Zhao, Cheng Peng, Zhao Yang, David V. Conti
}
