% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lucid_helper.R
\name{est_mu}
\alias{est_mu}
\title{M-step: obtain sparse mean for omics data via LASSO penalty}
\usage{
est_mu(j, rho, z, r, mu, wi)
}
\description{
M-step: obtain sparse mean for omics data via LASSO penalty
}
