% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lucid_helper.R
\name{Estep}
\alias{Estep}
\title{Calculate the log-likelihood of cluster assignment for each observation}
\usage{
Estep(
  beta,
  mu,
  sigma,
  gamma = NULL,
  G,
  Z,
  Y = NULL,
  family.list,
  K,
  N,
  useY,
  ind.na,
  ...
)
}
\description{
Calculate the log-likelihood of cluster assignment for each observation
}
