% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_estep.R
\name{f_XtoY}
\alias{f_XtoY}
\title{log f(Y|X)}
\usage{
f_XtoY(Y, Delta, family)
}
\arguments{
\item{Y}{an N by 1 matrix}

\item{Delta}{a list with parameters related to outcome}

\item{family}{a string, either gaussian or binomial}
}
\value{
for 2 omics layers, an K1 by K2 by N matrix; for 3 omics layers, an
K1 by K2 by K3 by N matrix
}
\description{
Calculate the log likelihood of outcome Y given all latent variables
X
}
