% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lucid.R
\name{fill_data_help1}
\alias{fill_data_help1}
\title{calculate the first half of the imputed values}
\usage{
fill_data_help1(obs, B, mu, alpha, sigma_inv, P)
}
\description{
calculate the first half of the imputed values
}
