% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lucid.R
\name{fill_data_help2}
\alias{fill_data_help2}
\title{calcualte the second half of the imputed values}
\usage{
fill_data_help2(obs, A, B, mu, alpha, sigma_inv, P)
}
\description{
calcualte the second half of the imputed values
}
