% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_lucid.R
\name{gen_ci}
\alias{gen_ci}
\title{generate bootstrp ci (normal, basic and percentile)}
\usage{
gen_ci(x, conf = 0.95)
}
\arguments{
\item{x}{an object return by boot function}

\item{conf}{A numeric scalar between 0 and 1 to specify confidence level(s)
of the required interval(s).}
}
\value{
a matrix, the first column is the point estimate from original model
}
\description{
generate bootstrp ci (normal, basic and percentile)
}
