% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lucid_helper.R
\name{gen_cov_matrices}
\alias{gen_cov_matrices}
\title{Use uniform distributon to initialzie variance covariance matrices}
\usage{
gen_cov_matrices(dimZ, K)
}
\description{
Use uniform distributon to initialzie variance covariance matrices
}
