% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{helix_data}
\alias{helix_data}
\title{HELIX data}
\format{
A list with 4 matrices corresponding to exposures (G), omics data (Z),
outcome (Y) and covariates (CoY)
\describe{
    \item{exposure}{18 variables, representing children/maternal exposure to 
    organochlorines}
    \item{omics}{35 untargeted metabolites and 10 proteins}
    \item{outcome}{A continuous outcome for BMI-z score based on WHO standard,
    A binary outcome for body mass index categories at 6-11 years old based on 
    WHO reference (0: Thinness or Normal; 1: Overweight or Obese)}
    \item{covariate}{5 covariates including mother's bmi, child sex, maternal
    age, study cohort and maternal education level}
}
}
\usage{
helix_data
}
\description{
The Human Early-Life Exposome (HELIX) project is multi-center 
research project that aims to characterize early-life environmental exposures
and associate these with omics biomarkers and child health outcomes (Vrijheid, 2014. doi: 10.1289/ehp.1307204).
We used a subset of HELIX data from Exposome Data Challenge 2021 (hold by
ISGlobal) as an example to illustrate LUCID model.
}
\keyword{datasets}
