% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{initialize_Beta}
\alias{initialize_Beta}
\title{Initialize Beta}
\usage{
initialize_Beta(K, nG)
}
\arguments{
\item{K}{A vector specifying the number of clusters for each omics layer.}

\item{nG}{The number of exposures.}
}
\description{
Initializes the Beta parameter for the EM algorithm.
}
