% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{initialize_Mu}
\alias{initialize_Mu}
\title{Initialize Mu}
\usage{
initialize_Mu(K, nZ)
}
\arguments{
\item{K}{A vector specifying the number of clusters for each omics layer.}

\item{nZ}{A vector specifying the number of variables for each omics layer.}
}
\description{
Initializes the Mu parameter for the EM algorithm.
}
