% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{Istep_Z}
\alias{Istep_Z}
\title{I-step of LUCID}
\usage{
Istep_Z(Z, p, mu, sigma, index, lucid_model)
}
\arguments{
\item{Z}{Matrix or list of matrices for omics data}

\item{p}{Probability matrix/vector or list for parallel model}

\item{mu}{Mean matrix or list of matrices}

\item{sigma}{Covariance array or list of arrays}

\item{index}{Missing value indicators}

\item{lucid_model}{Model type}
}
\value{
Imputed dataset
}
\description{
Impute missing data in Z by maximizing the likelihood given fixed parameters
}
