% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{LogSumExp}
\alias{LogSumExp}
\title{Log-Sum-Exp Trick}
\usage{
LogSumExp(vec)
}
\arguments{
\item{vec}{A numeric vector.}
}
\value{
The result of the log-sum-exp trick.
}
\description{
Computes the log-sum-exp trick for a vector.
}
