% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_lucid.R
\name{lucid_par}
\alias{lucid_par}
\title{function to calculate parameters of LUCID model. use as statisitc input for
boot function.}
\usage{
lucid_par(data, indices, model, dimG, dimZ, dimCoY, dimCoG)
}
\description{
function to calculate parameters of LUCID model. use as statisitc input for
boot function.
}
