% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lusc_data}
\alias{lusc_data}
\title{TCGA LUSC data}
\format{
A list with 3 matrices corresponding to exposure (G), 
omics measurements (Z) and a continuous outcome (survival time). 
\describe{
    \item{smoke_pack}{exposure, number of smoking packs per year}
    \item{methyl}{20 methylation measurements}
    \item{survtime}{survival time}
}
}
\source{
https://cran.r-project.org/web/packages/whitening/index.html
}
\usage{
lusc_data
}
\description{
This is a modified DNA methylation data combining selected 
clinical exposure (number of smoking packs per year) and outcome (survival 
time) for 130 patients with lung squamous cell carcinoma (LUSC). The original 
.RData file is extracted from R package 'whitening' (Strimmer et al. 2021), 
which is a pre-processed sample from The Cancer Genome Atlas (TCGA) database.
Please note: the missing values in smoking packs are imputed by random filling 
so the dataset is served as a illustration of LUCID analysis and shouldn't be
used for formal analysis.
}
\keyword{datasets}
