% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lucid_helper.R
\name{Mstep_G}
\alias{Mstep_G}
\title{Estiamte the association between exposure and latent cluster}
\usage{
Mstep_G(G, r, selectG, penalty, dimG, K)
}
\arguments{
\item{G}{}

\item{r}{}

\item{selectG}{}

\item{penalty}{}

\item{dimG}{}

\item{K}{}
}
\description{
Estiamte the association between exposure and latent cluster
}
