% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lucid.R
\name{Mstep_Z}
\alias{Mstep_Z}
\title{Title}
\usage{
Mstep_Z(Z, r, selectZ, penalty.mu, penalty.cov, model.name, K, ind.na, mu)
}
\arguments{
\item{Z}{}

\item{r}{}

\item{selectZ}{}

\item{penalty.mu}{}

\item{penalty.cov}{}

\item{model.name}{}

\item{K}{}

\item{ind.na}{}

\item{mu}{}
}
\value{

}
\description{
Title
}
