% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lucid_helper.R
\name{print.lucid}
\alias{print.lucid}
\title{Print the output of \code{est.lucid}}
\usage{
\method{print}{lucid}(x, ...)
}
\arguments{
\item{x}{An object of LUCID model, returned by \code{est.lucid}}

\item{...}{Other arguments to be passed to \code{print}}
}
\description{
Print the output of \code{est.lucid}
}
