% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_lucid.R
\name{print.sumlucid}
\alias{print.sumlucid}
\title{print a sumlucid object}
\usage{
\method{print}{sumlucid}(x)
}
\arguments{
\item{x}{A sumlucid object returned by \code{summary.lucid}}
}
\description{
print a sumlucid object
}
\examples{

}
