% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_utils.R
\name{safe_log_sum_exp}
\alias{safe_log_sum_exp}
\title{Safe log-sum-exp computation}
\usage{
safe_log_sum_exp(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numerically stable result
}
\description{
Safe log-sum-exp computation
}
