% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_utils.R
\name{safe_normalize}
\alias{safe_normalize}
\title{Safe probability normalization}
\usage{
safe_normalize(x)
}
\arguments{
\item{x}{Vector of probabilities}
}
\value{
Normalized probabilities
}
\description{
Safe probability normalization
}
