% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_HELIX_data}
\alias{simulated_HELIX_data}
\title{A simulated HELIX dataset for LUCID}
\format{
A list of 4 elements, each with 420 observations:
\describe{
\item{phenotype}{A 420 by 6 data frame holding the exposure, the outcome
and the covariates: \code{id}; \code{hs_hg_m_scaled}, the scaled
maternal exposure to in-utero mercury, used as \code{G};
\code{ck18_scaled}, a scaled continuous indicator of
metabolic-dysfunction-associated fatty liver disease (MAFLD), used as
\code{Y}; and the covariates \code{hs_child_age_yrs_None},
\code{e3_sex_None} and \code{h_fish_preg_Ter}.}
\item{methylome}{A 420 by 10 matrix of methylomics features.}
\item{transcriptome}{A 420 by 10 matrix of transcriptomics features.}
\item{miRNA}{A 420 by 10 matrix of miRNA features.}
}
}
\usage{
simulated_HELIX_data
}
\description{
The Human Early-Life Exposome (HELIX) project is a multi-center
research project that aims to characterize early-life environmental exposures
and associate these with omics biomarkers and child health outcomes (Vrijheid,
2014. doi: 10.1289/ehp.1307204). This is a simulated subset of the HELIX data
released for the Exposome Data Challenge 2021 (held by ISGlobal), used to
illustrate the LUCID model on three omics layers.

The three omics layers share the same 420 subjects in the same row order, so
they can be passed together as the \code{Z} list of a parallel or serial fit,
or used one at a time for early integration.
}
\keyword{datasets}
