% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{summarize_missing_stats}
\alias{summarize_missing_stats}
\title{Summarize missing-data patterns from check_na output}
\usage{
summarize_missing_stats(na_pattern, lucid_model = c("early", "parallel"))
}
\arguments{
\item{na_pattern}{Output list from check_na}

\item{lucid_model}{Model type ("early" or "parallel")}
}
\value{
Missing-data summary list
}
\description{
Summarize missing-data patterns from check_na output
}
