% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_lucid.R
\name{summary.lucid}
\alias{summary.lucid}
\title{Summarize the results of LUCID model}
\usage{
\method{summary}{lucid}(x, boot.se = NULL)
}
\arguments{
\item{x}{A model fitted by \code{\link{est.lucid}}}

\item{boot.se}{A object returned by \code{\link{boot.lucid}}, which contains the bootstrap standard error}
}
\value{
A list with class "sumlucid", which contains the following object
\item{Beta}{Estimates of genetic/environmental effects (and effect of covariates if included), matrix}
\item{Mu}{Estimates of cluster-specific biomarker means, matrix}
\item{Gamma}{Estimates of cluster-specific disease risk (and effect of covariates if included), vector}
\item{Family}{Type of Y, binary or normal}
\item{K}{Number of latent clusters}
\item{loglik}{log likelihood of the model}
\item{BIC}{Bayesian Information Criteria of the model}
\item{boot.se}{Bootstrap SE for estimates, an object returned by \code{\link{boot.lucid}}}
}
\description{
Summarize the results of LUCID model
}
\examples{

}
\references{
Cheng Peng, Jun Wang, Isaac Asante, Stan Louie, Ran Jin, Lida Chatzi, Graham Casey, Duncan C Thomas, David V Conti, A Latent Unknown Clustering Integrating Multi-Omics Data (LUCID) with Phenotypic Traits, Bioinformatics, , btz667, https://doi.org/10.1093/bioinformatics/btz667.
}
\author{
Yinqi Zhao, Cheng Peng, Zhao Yang, David V. Conti
}
