% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.lucid_parallel}
\alias{summary.lucid_parallel}
\title{Summarize results of the parallel LUCID model}
\usage{
\method{summary}{lucid_parallel}(object, ...)
}
\arguments{
\item{object}{A LUCID model fitted by \code{\link{estimate_lucid}}}

\item{...}{Additional argument \code{boot.se}, which can be an object
returned by \code{\link{boot_lucid}} to display bootstrap CIs in print output.}
}
\value{
A list containing model information, fit statistics, feature-selection
summaries (overall + by layer), detailed parameter estimates (by layer),
missing-data summaries (by layer), and optional bootstrap CI tables.
}
\description{
Summarize results of the parallel LUCID model
}
