% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_lucid}
\alias{summary_lucid}
\alias{summary.early_lucid}
\title{Summarize results of the early LUCID model}
\usage{
summary_lucid(object, ...)

\method{summary}{early_lucid}(object, ...)
}
\arguments{
\item{object}{A LUCID model fitted by \code{\link{estimate_lucid}}}

\item{...}{Additional argument \code{boot.se}, which can be an object
returned by \code{\link{boot_lucid}} to display bootstrap CIs in print output.}
}
\value{
A list containing model information, fit statistics, feature-selection
summaries, detailed parameter estimates, missing-data summaries, and optional
bootstrap CI tables.
}
\description{
Summarize results of the early LUCID model
}
\examples{
\donttest{
# use simulated data
G <- sim_data$G[1:300, , drop = FALSE]
Z <- sim_data$Z[1:300, , drop = FALSE]
Y_normal <- sim_data$Y_normal[1:300]

# fit lucid model
fit1 <- estimate_lucid(G = G, Z = Z, Y = Y_normal, lucid_model = "early", family = "normal", K = 2,
seed = 1008)

# conduct bootstrap resampling
boot1 <- suppressWarnings(
  boot_lucid(G = G, Z = Z, Y = Y_normal, lucid_model = "early", model = fit1, R = 5)
)

# summarize lucid model
summary(fit1)

# summarize lucid model with bootstrap CIs
summary(fit1, boot.se = boot1)
}
}
