% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y1}
\alias{Y1}
\title{Outcome Set 1}
\format{
A set with 3000 rows and 1 variable:
\describe{
  \item{Y1}{A binary outcome}
}
}
\usage{
Y1
}
\description{
A simulated dataset containing one of the components to run lucid. The variables are as follows:
}
\keyword{datasets}
