% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y2}
\alias{Y2}
\title{Outcome Set 2}
\format{
A set with 3000 rows and 1 variable:
\describe{
  \item{Y2}{A continuous outcome}
}
}
\usage{
Y2
}
\description{
A simulated dataset containing one of the components to run lucid. The variables are as follows:
}
\keyword{datasets}
