% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Z1}
\alias{Z1}
\title{Biomarker Set 1}
\format{
A set with 3000 rows and 4 variables:
\describe{
  \item{CZ1 - CZ5}{Causal biomarkers}
  \item{NZ1 - NZ5}{Null biomarkers}
}
}
\usage{
Z1
}
\description{
A simulated dataset containing one of the components to run est_lucid,
plot_lucid, and tune_lucid. The variables are as follows:
}
\keyword{datasets}
