% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Z2}
\alias{Z2}
\title{Biomarker Set 2}
\format{
A set with 3000 rows and 4 variables:
\describe{
  \item{CZ1 - CZ5}{Causal biomarkers}
  \item{NZ1 - NZ5}{Null biomarkers}
}
}
\usage{
Z2
}
\description{
A simulated dataset containing one of the components to run lucid. The variables are as follows:
}
\keyword{datasets}
