% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_lucid.R
\name{boot.lucid}
\alias{boot.lucid}
\title{Bootstrap}
\usage{
boot.lucid(G, Z, Y, CoG = NULL, CoY = NULL, model, R = 100)
}
\arguments{
\item{G}{}

\item{Z}{}

\item{Y}{}

\item{CoG}{}

\item{CoY}{}

\item{model}{}

\item{R}{}
}
\value{

}
\description{

}
