% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_lucid.R
\name{boot.lucid}
\alias{boot.lucid}
\title{Bootstrap}
\usage{
boot.lucid(G, Z, Y, CoG = NULL, CoY = NULL, model, R = 100)
}
\arguments{
\item{G}{Genetic features/environmental exposures, a \code{\link{matrix}}.}

\item{Z}{Biomarkers/other omics data, a \code{\link{matrix}}.}

\item{Y}{Disease outcome, it is suggested to transform it into a n by 1 \code{\link{matrix}}.}

\item{CoG}{Optional, matrix. Covariates to be adjusted for estimating the latent cluster.}

\item{CoY}{Optional, matrix. Covariates to be adjusted for estimating the outcome.}

\item{model}{A LUCID model fitted by \code{\link{est.lucid}}.}

\item{R}{Number of boostrap iterations.}
}
\value{

}
\description{
This function provides SEs of parameter estimates from a LUCID model through bootstrap method.
}
\examples{

}
\references{
Cheng Peng, Jun Wang, Isaac Asante, Stan Louie, Ran Jin, Lida Chatzi, Graham Casey, Duncan C Thomas, David V Conti, A Latent Unknown Clustering Integrating Multi-Omics Data (LUCID) with Phenotypic Traits, Bioinformatics, , btz667, https://doi.org/10.1093/bioinformatics/btz667.
}
\author{
Yinqi Zhao, Cheng Peng, Zhao Yang, David V. Conti
}
