% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{check_na}
\alias{check_na}
\title{Check missing patterns in omics data Z}
\usage{
check_na(Z)
}
\value{
1. index:indeces for missing values in omics data
2. indicator_na: missing pattern for each observation
3. impute_flag: - flag to initialize imputation. Only happens when sporadic missing
pattern is observed
}
\description{
Check missing patterns in omics data Z
}
