% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CovY}
\alias{CovY}
\title{Covariates Set 1}
\format{
A set with 3000 rows and 2 continuous variable:
\describe{
  \item{CovY1, CovY2}{Two continuous covariates}
}
}
\usage{
CovY
}
\description{
A simulated dataset containing covariates included in the X->Y analysis. The variables are as follows:
}
\keyword{datasets}
