% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_lucid.R
\name{def.tune}
\alias{def.tune}
\title{Define tuning parameters of regularization for LUCID model.}
\usage{
def.tune(
  Rho_G = 0,
  Rho_Z_InvCov = 0,
  Rho_Z_Mu = 0,
  Rho_Z_CovMu = 0,
  Select_G = FALSE,
  Select_Z = FALSE
)
}
\arguments{
\item{Rho_G}{Numeric. Penalty for selecting exposures G.}

\item{Rho_Z_InvCov}{Numeric. Penalty to obtain a sparse covariance matrix for omics data Z.}

\item{Rho_Z_Mu}{Numeric.}

\item{Rho_Z_CovMu}{Deprecated argument. Penalty for the product of the inverse of the covariance of omicss, which will produce a sparse matrix for the mean.}

\item{Select_G}{Deprecated argument. Flag for variable selection in genetic features/environmental exposures.}

\item{Select_Z}{Deprecated argument. Flag for variable selection in omics data Z.}
}
\description{
Define tuning parameters of regularization for LUCID model.
}
