% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lucid.R
\name{est.lucid}
\alias{est.lucid}
\title{Estimate latent unknown clusters with multi-omics data}
\usage{
est.lucid(
  G,
  Z,
  Y,
  CoG = NULL,
  CoY = NULL,
  K = 2,
  family = c("normal", "binary"),
  useY = TRUE,
  control = def.control(),
  tune = def.tune(),
  modelName = NULL,
  seed = 123,
  init_impute = c("mclust", "lod")
)
}
\arguments{
\item{G}{Exposures, a \code{matrix} or a \code{dataframe}.}

\item{Z}{Omics data, a \code{matrix} or a \code{dataframe}.}

\item{Y}{Outcome, an n x 1 \code{matrix}.}

\item{CoG}{Optional, a \code{matrix}. Covariates to be adjusted for estimating the latent cluster.}

\item{CoY}{Optional, a \code{matrix}. Covariates to be adjusted for estimating the association between latent cluster and outcome.}

\item{K}{The number of latent clusters. An integer greater or equal to 2. You can use \code{\link{tune.lucid}} to determine the optimal number of latent clusters.}

\item{family}{Distribution of outcome Y. Currently accepted distributions are "normal" and "binary".}

\item{useY}{Whether or not to include the information of Y to estimate the latent clusters. Default is TRUE.}

\item{control}{A list of tolerance parameters used by EM algorithm. See \code{\link{def.control}}.}

\item{tune}{A list of tuning parameters used by variable selection procedure. See \code{\link{def.tune}}}

\item{modelName}{The variance-covariance structure for omics data. See \code{\link{mclustModelNames}} for details.}

\item{seed}{An integer to initialize the EM algorithm}
}
\value{
A list which contains the several features of LUCID, including:
\item{pars}{Estimates of parameters of LUCID, including beta (estimates of genetic feature/environmental exposure), mu (estimates of cluster-specific biomarker means), sigma (estimates of the cluster-specific biomarker variance-covariance matrix) and gamma(estimates of cluster-specific effect and covariates effect related to the outcome)}
\item{K}{Number of latent cluster}
\item{modelName}{The model used to estimate the cluster-specific variance-covariance matrix, for further details, see \code{\link{mclust}}}
\item{likelihood}{The log likelihood of the LUCID model}
\item{post.p}{Predicted probability of belonging to each latent cluster}
\item{Z}{}
}
\description{
This function fits a LUCID model, which estimates the latent clusters by integrating genetic features/environmental exposures, biomarkers with/without the outcome of interest. Variable selection is available for analyzing the high-dimensional data.
}
\examples{
\dontrun{
set.seed(10)
fit1 <- est.lucid(G = sim1[, 1:10],Z = sim1[, 11:20],Y = as.matrix(sim1[, 21]), 
K = 2, family = "binary")
fit2 <- est.lucid(G = sim1[, 1:10],Z = sim1[, 11:20],Y = as.matrix(sim1[, 21]), 
K = 2, family = "binary", 
tune = def.tune(Select_Z = TRUE, Rho_Z_InvCov = 0.1, Rho_Z_CovMu = 90, 
Select_G = TRUE, Rho_G = 0.02))
}
}
\references{
Cheng Peng, Jun Wang, Isaac Asante, Stan Louie, Ran Jin, Lida Chatzi, Graham Casey, Duncan C Thomas, David V Conti, A Latent Unknown Clustering Integrating Multi-Omics Data (LUCID) with Phenotypic Traits, Bioinformatics, , btz667, https://doi.org/10.1093/bioinformatics/btz667.
}
\author{
Yinqi Zhao, Cheng Peng, Zhao Yang, David V. Conti
}
