% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lucid_helper.R
\name{est_mu}
\alias{est_mu}
\title{estimate the penalized mean in M-step Z}
\usage{
est_mu(j, rho, z, r, mu, wi)
}
\arguments{
\item{j}{}

\item{rho}{}

\item{z}{}

\item{r}{}

\item{mu}{}

\item{wi}{}
}
\description{
estimate the penalized mean in M-step Z
}
