% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lucid_helper.R
\name{Estep}
\alias{Estep}
\title{Calculate the log-likelihood of cluster assignment for each observation}
\usage{
Estep(
  beta,
  mu,
  sigma,
  gamma,
  G,
  Z,
  Y = NULL,
  family.list,
  K,
  N,
  useY,
  ind.na,
  ...
)
}
\arguments{
\item{beta}{}

\item{mu}{}

\item{sigma}{}

\item{gamma}{}

\item{G}{}

\item{Z}{}

\item{Y}{}

\item{family.list}{}

\item{K}{}

\item{N}{}

\item{useY}{}

\item{ind.na}{}
}
\value{

}
\description{
Calculate the log-likelihood of cluster assignment for each observation
}
