% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{fill_data_help1}
\alias{fill_data_help1}
\title{Calculate the first half of the imputed values}
\usage{
fill_data_help1(obs, B, mu, alpha, sigma_inv, P)
}
\description{
Calculate the first half of the imputed values
}
