% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_lucid.R
\name{gen_ci}
\alias{gen_ci}
\title{generate bootstrp ci (normal, basic and percentile)}
\usage{
gen_ci(x, conf = 0.95)
}
\arguments{
\item{x}{an object return by boot function}
}
\value{
a matrix, the first column is t0 statistic from original model
}
\description{
generate bootstrp ci (normal, basic and percentile)
}
