% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{Istep_Z}
\alias{Istep_Z}
\title{I-step of LUCID: impute missing data in Z by maximizing the likelihood}
\usage{
Istep_Z(Z, p, mu, sigma, index)
}
\value{
a complete dataset of Z
}
\description{
I-step of LUCID: impute missing data in Z by maximizing the likelihood
}
