% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lucid_helper.R
\name{lse_vec}
\alias{lse_vec}
\title{use the log-sum-exp trick to normalize a vector}
\usage{
lse_vec(vec)
}
\arguments{
\item{vec}{a vector of length K}
}
\description{
use the log-sum-exp trick to normalize a vector
}
