% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucid.R
\name{lucid}
\alias{lucid}
\title{A wrapper to estimate latent unknown clusters with multi-omics data}
\usage{
lucid(
  G,
  Z,
  Y,
  CoG = NULL,
  CoY = NULL,
  family = "normal",
  useY = TRUE,
  K = 2:5,
  Rho_G = 0,
  Rho_Z_InvCov = 0,
  Rho_Z_CovMu = 0
)
}
\arguments{
\item{G}{Required. Genetic features/environmental exposures, a \code{\link{matrix}}.}

\item{Z}{Required. Biomarkers/other omics data, a \code{\link{matrix}}.}

\item{Y}{Required. Outcome, it is suggested to transform it into a n by 1 \code{\link{matrix}}.}

\item{CoG}{Optional, matrix. Covariates to be adjusted for estimating the latent cluster.}

\item{CoY}{Optional, matrix. Covariates to be adjusted for estimating the outcome.}

\item{family}{Required. Type of outcome Y. It should be choose from "normal", "binary".}

\item{useY}{Whether or not to include the information of Y to estimate the latent clusters. Default is TRUE.}

\item{K}{Required, an array of integers, representing the number of latent clusters}

\item{Rho_G}{Optional, an array of penalties for variable selection in exposures}

\item{Rho_Z_InvCov}{Optional, an array of penalties for variable selection in Biomarkers (penalizing inverse of covariance matrix)}

\item{Rho_Z_CovMu}{Optional, an array of penalties for variable selection in Biomarkers (penalizing mean)}
}
\value{
A list:
\item{best_model}{the best model among different combination of tuning parameters}
\item{tune_list}{a data frame contains combination of tuning parameters and corresponding BIC}
\item{res_model}{a list stores LUCID model corresponding to each combination of tuning parameters}
}
\description{
A wrapper to estimate latent unknown clusters with multi-omics data
}
