% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucid.R
\name{lucid}
\alias{lucid}
\title{A wrapper function to perform model selection for LUCID}
\usage{
lucid(
  G,
  Z,
  Y,
  CoG = NULL,
  CoY = NULL,
  family = "normal",
  useY = TRUE,
  K = 2:5,
  Rho_G = 0,
  Rho_Z_Mu = 0,
  Rho_Z_Cov = 0,
  ...
)
}
\arguments{
\item{G}{Exposures, a numeric vector, matrix, or data frame. Categorical variable 
should be transformed into dummy variables. If a matrix or data frame, rows 
represent observations and columns correspond to variables.}

\item{Z}{Omics data, a numeric matrix or data frame. Rows correspond to observations
and columns correspond to variables.}

\item{Y}{Outcome, a numeric vector. Categorical variable is not allowed. Binary 
outcome should be coded as 0 and 1.}

\item{CoG}{Optional, covariates to be adjusted for estimating the latent cluster.
A numeric vector, matrix or data frame. Categorical variable should be transformed 
into dummy variables.}

\item{CoY}{Optional, covariates to be adjusted for estimating the association 
between latent cluster and the outcome. A numeric vector, matrix or data frame. 
Categorical variable should be transformed into dummy variables.}

\item{family}{Distribution of outcome. For continuous outcome, use "normal"; 
for binary outcome, use "binary". Default is "normal".}

\item{useY}{Flag to include information of outcome when estimating the latent 
cluster. Default is TRUE.}

\item{K}{Number of latent clusters. An integer greater or equal to 2.}

\item{Rho_G}{A scalar or a vector. Penalty to conduct LASSO regularization and 
obtain a sparse estimation for effect of exposures. If a vector, \code{lucid} will 
fit lucid model over the grid of penalties.}

\item{Rho_Z_Mu}{A scalar or a vector. Penalty to conduct LASSO regularization 
and obtain a sparse estimation of cluster-specific mean for omics data. If a 
vector, \code{lucid} will fit lucid model over the grid of penalties.}

\item{Rho_Z_Cov}{Penalty to conduct graphic LASSO regularization and obtain a
sparse estimation of cluster-specific variance-covariance matrices for omics 
data. If a vector, \code{lucid} will fit lucid model over the grid of penalties.}

\item{...}{Other parameters passed to \code{est.lucid}}
}
\value{
A list:
\item{best_model}{the best model over different combination of tuning parameters}
\item{tune_list}{a data frame contains combination of tuning parameters and c
orresponding BIC}
\item{res_model}{a list of LUCID models corresponding to each combination of 
tuning parameters}
}
\description{
Given a grid of K and L1 penalties (incluing Rho_G, Rho_Z_mu and
Rho_Z_Cov), fit LUCID model over all combinations of K and L1 penalties to 
determine the optimal penalty.
}
\examples{
\dontrun{

}
}
