% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lucid_helper.R
\name{Mstep_G}
\alias{Mstep_G}
\title{M-step to estimate the association between exposure and latent cluster}
\usage{
Mstep_G(G, r, selectG, penalty, dimG, dimCoG, K)
}
\description{
M-step to estimate the association between exposure and latent cluster
}
