% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lucid.R
\name{Mstep_G}
\alias{Mstep_G}
\title{Title}
\usage{
Mstep_G(G, r, selectG, penalty, dimG, K)
}
\arguments{
\item{G}{}

\item{r}{}

\item{selectG}{}

\item{penalty}{}

\item{dimG}{}

\item{K}{}
}
\value{

}
\description{
Title
}
