% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lucid_helper.R
\name{Mstep_Z}
\alias{Mstep_Z}
\title{M-step to estimate the parameters related to GMM for omics data}
\usage{
Mstep_Z(Z, r, selectZ, penalty.mu, penalty.cov, model.name, K, ind.na, mu)
}
\description{
M-step to estimate the parameters related to GMM for omics data
}
