% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lucid.R
\name{plot.lucid_serial}
\alias{plot.lucid_serial}
\title{Visualize LUCID model through a Sankey diagram}
\usage{
\method{plot}{lucid_serial}(
  x,
  G_color = "dimgray",
  X_color = "#eb8c30",
  Z_color = "#2fa4da",
  Y_color = "#afa58e",
  pos_link_color = "#67928b",
  neg_link_color = "#d1e5eb",
  fontsize = 7
)
}
\arguments{
\item{x}{A LUCID model fitted by \code{\link{estimate_lucid}}}

\item{G_color}{Color of node for exposure}

\item{X_color}{Color of node for latent cluster}

\item{Z_color}{Color of node for omics data}

\item{Y_color}{Color of node for outcome}

\item{pos_link_color}{Color of link corresponds to positive association}

\item{neg_link_color}{Color of link corresponds to negative association}

\item{fontsize}{Font size for annotation}
}
\value{
A DAG graph created by \code{\link{sankeyNetwork}}
}
\description{
In the Sankey diagram, each node either represents a variable (exposure,
omics or outcome) or a latent cluster. Each line represents an association. The
color of the node represents variable type, either exposure, omics or outcome.
The width of the line represents the effect size of a certain association; the
color of the line represents the direction of a certain association. Only work for LUCID early for now.
}
