% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_lucid.R
\name{predict_lucid}
\alias{predict_lucid}
\title{Predict cluster assignment and outcome based on LUCID model}
\usage{
predict_lucid(model, G, Z, CoG = NULL, CoY = NULL, response = TRUE, ...)
}
\arguments{
\item{model}{A model fitted and returned by \code{\link{est.lucid}}}

\item{G}{A new data set of genetic/environmental factors}

\item{Z}{A new data set of biomarkers}

\item{CoG}{Optional. A new data set of covariates included in the G->X analysis}

\item{CoY}{Optional. A new data set of covariates included in the X->Y analysis}

\item{response}{Report the posterior distribution of cluster assignment (and the probability of binary outcome), default is TRUE}

\item{...}{Other parameters to be passed to \code{predict}}
}
\value{
A list contains predicted latent cluster and outcome for each observation
}
\description{
Predict cluster assignment and outcome based on LUCID model
}
\examples{
\dontrun{
# prepare data
G <- sim_data$G
Z <- sim_data$Z
Y_normal <- sim_data$Y_normal

# fit lucid model
fit1 <- est.lucid(G = G, Z = Z, Y = Y_normal, K = 2, family = "normal")

# prediction on training set
pred1 <- predict_lucid(model = fit1, G = G, Z = Z)
}
}
