% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lucid.R
\name{print.lucid}
\alias{print.lucid}
\title{print a lucid object}
\usage{
\method{print}{lucid}(x)
}
\arguments{
\item{x}{fitted \code{lucid} object}
}
\description{
Print a summary of the results for a LUCID model.
}
\examples{
set.seed(10)
fit1 <- est.lucid(G = G1,Z = Z1,Y = Y1, CoY = CovY, K = 2, family = "binary")
print(fit1)
}
