% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_lucid.R
\name{print.sumlucid}
\alias{print.sumlucid}
\title{Print the output of LUCID in a nicer table}
\usage{
\method{print}{sumlucid}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{summary_lucid}}

\item{...}{Other parameters to be passed to \code{print}}
}
\description{
Print the output of LUCID in a nicer table
}
