% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.sumlucid_serial}
\alias{print.sumlucid_serial}
\title{Print the output of LUCID in a nicer table}
\usage{
\method{print}{sumlucid_serial}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{summary}}

\item{...}{Other parameters to be passed to \code{print.sumlucid_serial}}
}
\value{
A nice table/several nice tables of the summary of the LUCID model
}
\description{
Print the output of LUCID in a nicer table
}
