% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{A simulated dataset for LUCID}
\format{
A list with 5 matrices corresponding to exposures (G), omics data (Z),
a continuous outcome, a binary outcome and 2 covariates (can be used either 
as CoX or CoY). Each matrice contains 2000 observations.
\describe{
    \item{G}{10 exposures}
    \item{Z}{10 metabolites}
    \item{Y_normal}{Outcome, PFAS concentration in children}
    \item{Y_bninary}{Bianry outcome, liver injury status}
    \item{Covariates}{2 continous covariates, can be treated as either CoX or 
                      CoY}
    \item{X}{Latent clusters}
}
}
\usage{
sim_data
}
\description{
This is an example dataset to illustrate LUCID model. It is simulated
by assuming there are 2 latent clusters in the data. We assume the exposures 
are associated with latent cluster which ultimately affects the PFAS concentration
and liver injury in children. The latent clusters are also characterized by 
differential levels of metabolites.
}
\keyword{datasets}
